Refactored receipt functions to separate util class
